feat: export migration scripts (#113)
* export migration scripts

* export DbConfig from data-model

Export models
