+++ packages/data-model/src/config/db.ts
@@ -1,0 +1,7 @@
+export type DbConfig = {
+	host: string;
+	port: number;
+	database: string;
+	user: string;
+	password: string;
+};
