+++ packages/data-provider/src/config/config.ts
@@ -1,19 +1,12 @@
 import { NodePgDatabase } from 'drizzle-orm/node-postgres';
 
-import * as schema from '@overture-stack/lyric-data-model';
+import type { DbConfig } from '@overture-stack/lyric-data-model';
+import * as schema from '@overture-stack/lyric-data-model/models';
 
 import { Logger } from './logger.js';
 
 export type AuditConfig = {
 	enabled: boolean;
-};
-
-export type DbConfig = {
-	host: string;
-	port: number;
-	database: string;
-	user: string;
-	password: string;
 };
 
 export type RecordHierarchyConfig = {
