+++ packages/data-provider/src/repository/categoryRepository.ts
@@ -2,7 +2,7 @@ import { eq } from 'drizzle-orm/sql';
 import { ListAllCategoriesResponse } from 'src/utils/types.js';
 
 import { Dictionary as SchemasDictionary } from '@overture-stack/lectern-client';
-import { Category, Dictionary, dictionaryCategories, NewCategory } from '@overture-stack/lyric-data-model';
+import { Category, Dictionary, dictionaryCategories, NewCategory } from '@overture-stack/lyric-data-model/models';
 
 import { BaseDependencies } from '../config/config.js';
 import { ServiceUnavailable } from '../utils/errors.js';
