+++ packages/data-provider/src/utils/submittedDataUtils.ts
@@ -11,7 +11,7 @@ import {
 	type SubmissionDeleteData,
 	type SubmissionUpdateData,
 	SubmittedData,
-} from '@overture-stack/lyric-data-model';
+} from '@overture-stack/lyric-data-model/models';
 
 import {
 	DataRecordReference,
